250105 찜 CRD 완료

+++ frontserver/src/components/orders/userWishes.vue
@@ -11,8 +11,8 @@
                 </tr>
             </thead>
             <tbody>
-                <tr v-for="product in userWishList" :key="product.id">
-                    <td><img :src="product.product_description_img"></td>
+                <tr v-for="product in wishedProducts" :key="product.id">
+                    <td><img :src="product.product_image"></td>
                     <td>{{product.product_name}}</td>
                     <td>{{product.product_price}}</td>
                     <td><button @click="deleteProduct(product)">삭제</button></td>
@@ -30,7 +30,7 @@ export default{
     components:{},
     data(){
         return{
-            userWishList:[],
+            wishedProducts:[],
             userid: 0,
         };
     },
@@ -45,11 +45,12 @@ export default{
             try{
                 //const userId = this.session.userId;
                 this.userid = this.$route.params.userId;
-                console.log(`################프론트userId:${userId}`);
+                console.log(`################프론트userid:${this.userid}`);
                 // 로그인된 유저가 가지고있는 찜 리스트만 불러온다. 
-                const response = await axios.get(`http://localhost:3000/orders/wish/${userId}`); 
+                const response = await axios.get(`http://localhost:3000/orders/wish/${this.userid}`); 
                                                                         // get 요청보낼때는 data 말고 params로 보내자
-                this.userWishList = response.data;
+                this.wishedProducts = response.data;
+                console.log(this.wishedProducts)
             } catch(err){
                 console.error(err);
             }
@@ -57,18 +58,25 @@ export default{
 
         async deleteProduct(product) {
             try{
+                const deleteWish = confirm("해당 제품을 찜에서 삭제하실거에요?");
+                
+                //'취소' 누르면 함수 바로 종료
+                if(!deleteWish){return;}
+                console.log(`################Product선택된거:${JSON.stringify(product.id)}`);
                 const response = await axios.delete(`http://localhost:3000/orders/wish/${product.id}`)
+
+                // 삭제에 성공하면 위시리스트 다시 받아오지 말고 그냥 기존의 위시리스트에서 삭제항목만 빼고 보여준다.       
                 if (response.status === 200) {
-                    // 삭제에 성공하면 위시리스트 다시 받아오지 말고 그냥 기존의 위시리스트에서 삭제항목만 빼고 보여준다.
-                    this.userWishList = this.userWishList.filter(w => w.id !== product.id);
+                    alert("찜에서 삭제되었다");
+                    this.wishedProducts = this.wishedProducts.filter(w => w.id !== product.id);
                 }
                 console.log(response);
+
             }catch(error){
                 alert("찜 삭제 실패");
                 console.log("찜 삭제 실패", error);
             }
         }
-
     }
 }
